drivers: net: Fix void pointer to enum cast ../drivers/net/can/spi/mcp25xxfd.c:4266:17: error: cast to smaller integer type 'enum mcp25xxfd_model' from 'const void *' [-Werror,-Wvoid-pointer-to-enum-cast]                 priv->model = (enum mcp25xxfd_model)of_id->data;                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 1 error generated.

+++ drivers/net/can/spi/mcp25xxfd.c
@@ -4263,7 +4263,7 @@ static int mcp25xxfd_can_probe(struct spi_device *spi)
 		CAN_CTRLMODE_ONE_SHOT;
 
 	if (of_id)
-		priv->model = (enum mcp25xxfd_model)of_id->data;
+		priv->model = (enum mcp25xxfd_model)(long)of_id->data;
 	else
 		priv->model = spi_get_device_id(spi)->driver_data;
